==> 01rscripts/03deeplearning/06_training_config/01_train_control.R <==
#*******************************************************************************
#*******************************************************************************
#######################    train control  ######################################
#*******************************************************************************
#*******************************************************************************

# config train control

#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# define validation method
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

valid_method <- "repeatedcv"
valid_folds <- 2
valid_repeats <- 1

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# seeds
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
set.seed(42)
seeds <- setSeedsRepeatedCV(numbers = valid_folds, 
                            repeats = valid_repeats, 
                            tunes = nrow(tune_grid))

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# define training control
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

ctrl <- trainControl(
  method = valid_method, number = valid_folds, repeats = valid_repeats,
  summaryFunction = twoClassSummaryCustom,
  classProbs = TRUE,
  savePredictions = TRUE,
  allowParallel = TRUE,
  verboseIter = TRUE,
  seeds = seeds,
  selectionFunction = "best"
)
